first part of some industry rework done
