+++ angelsindustries/data-updates.lua
@@ -2,5 +2,6 @@ local OV = angelsmods.functions.OV
 
 require("prototypes.buildings.nuclear-reactor-updates")
 require("prototypes.angels-industries-override")
+require("prototypes.angels-industries-ordening")
 
 OV.execute()
