+++ angelsindustries/prototypes/angels-industries-ordening.lua
@@ -1,0 +1,7 @@
+-- ordening
+require("prototypes.ordening.angels-industries-override-ordening") -- TEMPORARY, FILE HAS TO GO *POOF*
+
+require("prototypes.ordening.angels-vehicles")
+require("prototypes.ordening.angels-logistics")
+require("prototypes.ordening.angels-power")
+require("prototypes.ordening.angels-intermediates")
